upgrade libraries and fix the db script

+++ scripts/init_db.sh
@@ -10,7 +10,7 @@ fi
 if ! [ -x "$(command -v sqlx)" ]; then
   echo >&2 "Error: sqlx is not installed."
   echo >&2 "Use:"
-  echo >&2 " cargo install --version=0.5.13 sqlx-cli --no-default-features --features postgres"
+  echo >&2 " cargo install --version=0.6 sqlx-cli --no-default-features --features postgres"
   echo >&2 "to install it."
   exit 1
 fi
@@ -19,8 +19,9 @@ DB_USER=${POSTGRES_USER:=postgres}
 DB_PASSWORD="${POSTGRES_PASSWORD:=password}"
 DB_NAME="${POSTGRES_DB:=newsletter-signup-service}"
 DB_PORT="${POSTGRES_PORT:=5432}"
+export PGPASSWORD="${DB_PASSWORD}"
 # Launch postgres using Docker
-if [[ -z "${SKIP_DOCKER}" ]]
+if [[ -z "${SKIP_DOCKER}" && $(psql -h "localhost" -U "${DB_USER}" -p "${DB_PORT}" -d "postgres" -c '\q';) ]]
 then
   docker system prune --volumes --force
 
@@ -35,7 +36,6 @@ fi
 # ˆ Increased maximum number of connections for testing purposes
 
 # Keep pinging Postgres until it's ready to accept commands
-export PGPASSWORD="${DB_PASSWORD}"
 until psql -h "localhost" -U "${DB_USER}" -p "${DB_PORT}" -d "postgres" -c '\q'; do
   >&2 echo "Postgres is still unavailable - sleeping"
 sleep 1
